Added ability to enter cities without accents
Gaspé was not working properly before, added check to see if cityName == cityAscii
to allow this. Many general clean ups. Removed an unneccessary try catch in client
set province call since it is handled in the statistics method. One small issue I ran
into is that updating the population replaces the &apos's in the XML with an actual
apostrophe but I'm not going to consider that wrong